Refactored the exercise and did a few changes to make the exercise as modular as possible. Added comments.
The most significant change:
The `response_=` method is now `respond` and it takes a cancellation
token, in case the response is very long.

Will comment more in an e-mail.

Todo: better integrate `Future.delay`, `continue`, `continueWith`,
`all`, `any`, etc., because they are not used right now at all.

Todo: Have an example with a server with an infinite response,
to show that it correctly terminates.